Move the config test to sqlbk
