+++ lib/backend/postgres/backend_test.go
@@ -159,120 +159,6 @@ storage:
 	require.Equal(t, expect, cfg)
 }
 
-func TestConfigCheck(t *testing.T) {
-	t.Parallel()
-
-	cases := []struct {
-		name string
-		cfg  *Config
-		f    require.ErrorAssertionFunc
-	}{
-		{
-			name: "TLS ok",
-			cfg: &Config{
-				Config: sqlbk.Config{
-					Addr: "-",
-					TLS: sqlbk.TLSAuthConfig{
-						ClientKeyFile:  "-",
-						ClientCertFile: "-",
-						CAFile:         "-",
-					},
-				},
-			},
-			f: require.NoError,
-		},
-		{
-			name: "TLS missing parameter",
-			cfg: &Config{
-				Config: sqlbk.Config{
-					Addr: "-",
-					TLS: sqlbk.TLSAuthConfig{
-						ClientKeyFile:  "-",
-						ClientCertFile: "",
-						CAFile:         "-",
-					},
-				},
-			},
-			f: require.Error,
-		},
-		{
-			name: "Azure ok",
-			cfg: &Config{
-				Config: sqlbk.Config{
-					Addr: "-",
-					TLS: sqlbk.TLSAuthConfig{
-						ClientKeyFile:  "",
-						ClientCertFile: "",
-						CAFile:         "-",
-					},
-					Azure: sqlbk.AzureAuthConfig{
-						Username: "-",
-					},
-				},
-			},
-			f: require.NoError,
-		},
-		{
-			name: "TLS Azure conflict",
-			cfg: &Config{
-				Config: sqlbk.Config{
-					Addr: "-",
-					TLS: sqlbk.TLSAuthConfig{
-						ClientKeyFile:  "-",
-						ClientCertFile: "-",
-						CAFile:         "-",
-					},
-					Azure: sqlbk.AzureAuthConfig{
-						Username: "-",
-					},
-				},
-			},
-			f: require.Error,
-		},
-		{
-			name: "Double username",
-			cfg: &Config{
-				Config: sqlbk.Config{
-					Addr: "-",
-					TLS: sqlbk.TLSAuthConfig{
-						CAFile:   "-",
-						Username: "-",
-					},
-					Azure: sqlbk.AzureAuthConfig{
-						Username: "-",
-					},
-				},
-			},
-			f: require.Error,
-		},
-		{
-			name: "TLS with Azure client ID",
-			cfg: &Config{
-				Config: sqlbk.Config{
-					Addr: "-",
-					TLS: sqlbk.TLSAuthConfig{
-						ClientKeyFile:  "-",
-						ClientCertFile: "-",
-						CAFile:         "-",
-						Username:       "-",
-					},
-					Azure: sqlbk.AzureAuthConfig{
-						Username: "",
-						ClientID: "-",
-					},
-				},
-			},
-			f: require.Error,
-		},
-	}
-
-	for _, tt := range cases {
-		t.Run(tt.name, func(t *testing.T) {
-			tt.f(t, tt.cfg.CheckAndSetDefaults())
-		})
-	}
-}
-
 // TestDriverURL verifies the correct connection string URL
 // is created from a Config.
 func TestDriverURL(t *testing.T) {
